implemented review points
